Revert "Extract evaluation procedure from training extension `Evaluator`"
