+++ chainer/training/__init__.py
@@ -1,8 +1,8 @@
-from chainer.training import evaluator
 from chainer.training import extension
 from chainer.training import trainer
 from chainer.training import trigger
 from chainer.training import updater
+
 
 Extension = extension.Extension
 make_extension = extension.make_extension
@@ -18,6 +18,3 @@ get_trigger = trigger.get_trigger
 Updater = updater.Updater
 StandardUpdater = updater.StandardUpdater
 ParallelUpdater = updater.ParallelUpdater
-
-Evaluator = evaluator.Evaluator
-StandardEvaluator = evaluator.StandardEvaluator
